Convert `statuses` spec controller->system/request
